combine unit tests
